feat: criando arquivo de types e exportando o type de video
