+++ client/front/src/types/videos.ts
@@ -1,0 +1,6 @@
+export type Videos = {
+  title: string;
+  _id: string;
+  link: string;
+  like: boolean;
+};
